moved api endpoints from `config` to `server_eps`
